requirements.txt
python-telegram-bot
python-dotenv
flask
psycopg2-binary
